+++ sources/treat_charflags.c
@@ -1,19 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   str_flags.c                                        :+:      :+:    :+:   */
+/*   treat_charflags.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: hanmpark <hanmpark@student.42nice.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/29 17:43:57 by hanmpark          #+#    #+#             */
-/*   Updated: 2023/01/02 12:51:11 by hanmpark         ###   ########.fr       */
+/*   Updated: 2023/01/03 16:41:07 by hanmpark         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/ft_printf.h"
-#include <stdio.h>
 
-void	left_justifystr(t_parseflags *tab, char *str, size_t len)
+void	left_justifystr(t_parseflags *tab, char *str, int len)
 {
 	int	toprint;
 
@@ -33,17 +32,17 @@ void	left_justifyc(t_parseflags *tab, char c)
 		tab->len += write(1, " ", 1);
 }
 
-void	precision_str(t_parseflags *tab, char *str, size_t len)
+void	precision_str(t_parseflags *tab, char *str, int len)
 {
 	while (*str && len--)
 		tab->len += write(1, &*str++, 1);
 }
 
-void	width(t_parseflags *tab, char *str, size_t len)
+void	width(t_parseflags *tab, char *str, int len)
 {
 	int	toprint;
 
-	toprint = tab->width - (int)len;
+	toprint = tab->width - len;
 	while (toprint-- > 0)
 		tab->len += write(1, " ", 1);
 	precision_str(tab, str, len);
